Refactoring Os specific packages implementation
