+++ src/main/java/com/sparetimelabs/serial/termios/impl/FreeBSDTermios.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, Kustaa Nyholm / SpareTimeLabs
+ * Copyright (c) 2012 Kustaa Nyholm / SpareTimeLabs
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, 
@@ -26,27 +26,26 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
+ * 
+ * This FreeBSD backend contributed by Denver Hull 
+ * 
+ * Many thanks for his persistence and efforts to make it happen!
+ * 
  */
-package com.sparetimelabs.serial.termios.linux;
+package com.sparetimelabs.serial.termios.impl;
 
 import com.sun.jna.*;
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStreamReader;
-import java.io.IOException;
 
 import java.util.*;
-import java.util.regex.Pattern;
 
-import com.sparetimelabs.serial.termios.JTermios;
 import com.sparetimelabs.serial.termios.Pollfd;
 import com.sparetimelabs.serial.termios.Termios;
 import com.sparetimelabs.serial.termios.TimeVal;
 import static com.sparetimelabs.serial.termios.JTermios.*;
 import static com.sparetimelabs.serial.termios.JTermios.JTermiosLogging.log;
 
-public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.JTermiosInterface {
+public class FreeBSDTermios implements com.sparetimelabs.serial.termios.JTermios.JTermiosInterface {
 
     private static String DEVICE_DIR_PATH = "/dev/";
     static C_lib_DirectMapping m_ClibDM;
@@ -61,46 +60,6 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         m_Clib = m_ClibDM;
     }
 
-    private final static int TIOCGSERIAL = 0x0000541E;
-    private final static int TIOCSSERIAL = 0x0000541F;
-
-    private final static int ASYNC_SPD_MASK = 0x00001030;
-    private final static int ASYNC_SPD_CUST = 0x00000030;
-
-    private final static int[] m_BaudRates
-            = { //
-                50, 0000001, //
-                75, 0000002, //
-                110, 0000003, //
-                134, 0000004, //
-                150, 0000005, //
-                200, 0000006, //
-                300, 0000007, //
-                600, 0000010, //
-                1200, 0000011, //
-                1800, 0000012, //
-                2400, 0000013, //
-                4800, 0000014, //
-                9600, 0000015, //
-                19200, 0000016, //
-                38400, 0000017, //
-                57600, 0010001, //
-                115200, 0010002, //
-                230400, 0010003, //
-                460800, 0010004, //
-                500000, 0010005, //
-                576000, 0010006, //
-                921600, 0010007, //
-                1000000, 0010010, //
-                1152000, 0010011, //
-                1500000, 0010012, //
-                2000000, 0010013, //
-                2500000, 0010014, //
-                3000000, 0010015, //
-                3500000, 0010016, //
-                4000000, 0010017 //
-            };
-
     public static class C_lib_DirectMapping implements C_lib {
 
         native public int pipe(int[] fds);
@@ -113,8 +72,6 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         native public int ioctl(int fd, int cmd, int[] arg);
 
-        native public int ioctl(int fd, int cmd, serial_struct arg);
-
         native public int open(String path, int flags);
 
         native public int close(int fd);
@@ -123,13 +80,13 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         native public int tcsetattr(int fd, int cmd, termios termios);
 
-        native public int cfsetispeed(termios termios, int i);
+        native public int cfsetispeed(termios termios, NativeLong i);
 
-        native public int cfsetospeed(termios termios, int i);
+        native public int cfsetospeed(termios termios, NativeLong i);
 
-        native public int cfgetispeed(termios termios);
+        native public NativeLong cfgetispeed(termios termios);
 
-        native public int cfgetospeed(termios termios);
+        native public NativeLong cfgetospeed(termios termios);
 
         native public NativeSize write(int fd, byte[] buffer, NativeSize count);
 
@@ -140,6 +97,9 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         native public void perror(String msg);
 
         native public int tcsendbreak(int fd, int duration);
+
+        native public int select(int n, fd_set read, fd_set write, fd_set error, timeval timeout);
+
     }
 
     public interface C_lib extends com.sun.jna.Library {
@@ -154,8 +114,6 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public int ioctl(int fd, int cmd, int[] arg);
 
-        public int ioctl(int fd, int cmd, serial_struct arg);
-
         public int open(String path, int flags);
 
         public int close(int fd);
@@ -164,13 +122,13 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public int tcsetattr(int fd, int cmd, termios termios);
 
-        public int cfsetispeed(termios termios, int i);
+        public int cfsetispeed(termios termios, NativeLong i);
 
-        public int cfsetospeed(termios termios, int i);
+        public int cfsetospeed(termios termios, NativeLong i);
 
-        public int cfgetispeed(termios termios);
+        public NativeLong cfgetispeed(termios termios);
 
-        public int cfgetospeed(termios termios);
+        public NativeLong cfgetospeed(termios termios);
 
         public NativeSize write(int fd, byte[] buffer, NativeSize count);
 
@@ -182,11 +140,11 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public int tcsendbreak(int fd, int duration);
 
+        public int select(int n, fd_set read, fd_set write, fd_set error, timeval timeout);
+
     }
 
     public interface NonDirectCLib extends com.sun.jna.Library {
-
-        public int select(int n, fd_set read, fd_set write, fd_set error, timeval timeout);
 
         public int poll(pollfd.ByReference pfds, int nfds, int timeout);
     }
@@ -239,15 +197,9 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
     static public class fd_set extends Structure implements FDSet {
 
-        private final static int NFBBITS = NativeLong.SIZE * 8;
+        private final static int NFBBITS = 32;
         private final static int fd_count = 1024;
-        public NativeLong[] fd_array = new NativeLong[(fd_count + NFBBITS - 1) / NFBBITS];
-
-        public fd_set() {
-            for (int i = 0; i < fd_array.length; ++i) {
-                fd_array[i] = new NativeLong();
-            }
-        }
+        public int[] fd_array = new int[(fd_count + NFBBITS - 1) / NFBBITS];
 
         @Override
         protected List getFieldOrder() {
@@ -257,72 +209,22 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         }
 
         public void FD_SET(int fd) {
-            fd_array[fd / NFBBITS].setValue(fd_array[fd / NFBBITS].longValue() | (1L << (fd % NFBBITS)));
+            fd_array[fd / NFBBITS] |= (1 << (fd % NFBBITS));
         }
 
         public boolean FD_ISSET(int fd) {
-            return (fd_array[fd / NFBBITS].longValue() & (1L << (fd % NFBBITS))) != 0;
+            return (fd_array[fd / NFBBITS] & (1 << (fd % NFBBITS))) != 0;
         }
 
         public void FD_ZERO() {
-            for (NativeLong fd : fd_array) {
-                fd.setValue(0L);
-            }
+            Arrays.fill(fd_array, 0);
         }
 
         public void FD_CLR(int fd) {
-            fd_array[fd / NFBBITS].setValue(fd_array[fd / NFBBITS].longValue() & ~(1L << (fd % NFBBITS)));
+            fd_array[fd / NFBBITS] &= ~(1 << (fd % NFBBITS));
         }
 
     }
-
-    public static class serial_struct extends Structure {
-
-        public int type;
-        public int line;
-        public int port;
-        public int irq;
-        public int flags;
-        public int xmit_fifo_size;
-        public int custom_divisor;
-        public int baud_base;
-        public short close_delay;
-        public short io_type;
-        //public char io_type;
-        //public char reserved_char;
-        public int hub6;
-        public short closing_wait;
-        public short closing_wait2;
-        public Pointer iomem_base;
-        public short iomem_reg_shift;
-        public int port_high;
-        public NativeLong iomap_base;
-
-        @Override
-        protected List getFieldOrder() {
-            return Arrays.asList(//
-                    "type",//
-                    "line",//
-                    "port",//
-                    "irq",//
-                    "flags",//
-                    "xmit_fifo_size",//
-                    "custom_divisor",//
-                    "baud_base",//
-                    "close_delay",//
-                    "io_type",//
-                    //public char io_type;
-                    //public char reserved_char;
-                    "hub6",//
-                    "closing_wait",//
-                    "closing_wait2",//
-                    "iomem_base",//
-                    "iomem_reg_shift",//
-                    "port_high",//
-                    "iomap_base"//
-            );
-        }
-    };
 
     static public class termios extends Structure {
 
@@ -330,8 +232,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         public int c_oflag;
         public int c_cflag;
         public int c_lflag;
-        public byte c_line;
-        public byte[] c_cc = new byte[32];
+        public byte[] c_cc = new byte[20];
         public int c_ispeed;
         public int c_ospeed;
 
@@ -342,7 +243,6 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
                     "c_oflag",//
                     "c_cflag",//
                     "c_lflag",//
-                    "c_line",//
                     "c_cc",//
                     "c_ispeed",//
                     "c_ospeed"//
@@ -356,8 +256,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
             c_iflag = t.c_iflag;
             c_oflag = t.c_oflag;
             c_cflag = t.c_cflag;
-            c_lflag = t.c_lflag;
-            System.arraycopy(t.c_cc, 0, c_cc, 0, t.c_cc.length);
+            System.arraycopy(t.c_cc, 0, c_cc, 0, Math.min(t.c_cc.length, c_cc.length));
             c_ispeed = t.c_ispeed;
             c_ospeed = t.c_ospeed;
         }
@@ -366,121 +265,124 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
             t.c_iflag = c_iflag;
             t.c_oflag = c_oflag;
             t.c_cflag = c_cflag;
-            t.c_lflag = c_lflag;
-            System.arraycopy(c_cc, 0, t.c_cc, 0, t.c_cc.length);
+            System.arraycopy(c_cc, 0, t.c_cc, 0, Math.min(t.c_cc.length, c_cc.length));
             t.c_ispeed = c_ispeed;
             t.c_ospeed = c_ospeed;
         }
     }
 
-    public JTermiosImpl() {
+    public FreeBSDTermios() {
         log = log && log(1, "instantiating %s\n", getClass().getCanonicalName());
 
-        //linux/serial.h stuff
-        FIONREAD = 0x541B; // Looked up manually
-        //fcntl.h stuff
+        // sys/filio.h stuff
+        FIONREAD = 0x4004667F;
+        // fcntl.h stuff
         O_RDWR = 0x00000002;
-        O_NONBLOCK = 0x00000800;
-        O_NOCTTY = 0x00000100;
-        O_NDELAY = 0x00000800;
+        O_NONBLOCK = 0x00000004;
+        O_NOCTTY = 0x00008000;
+        O_NDELAY = 0x00000004;
         F_GETFL = 0x00000003;
         F_SETFL = 0x00000004;
-        //errno.h stuff
-        EAGAIN = 11;
-        EACCES = 13;
+        // errno.h stuff
+        EAGAIN = 35;
+        EBADF = 9;
+        EACCES = 22;
         EEXIST = 17;
         EINTR = 4;
         EINVAL = 22;
         EIO = 5;
         EISDIR = 21;
-        ELOOP = 40;
+        ELOOP = 62;
         EMFILE = 24;
-        ENAMETOOLONG = 36;
+        ENAMETOOLONG = 63;
         ENFILE = 23;
         ENOENT = 2;
-        ENOSR = 63;
         ENOSPC = 28;
         ENOTDIR = 20;
         ENXIO = 6;
-        EOVERFLOW = 75;
+        EOVERFLOW = 84;
         EROFS = 30;
-        ENOTSUP = 95;
-        //termios.h stuff
+        ENOTSUP = 45;
+        // termios.h stuff
         TIOCM_RNG = 0x00000080;
         TIOCM_CAR = 0x00000040;
         IGNBRK = 0x00000001;
         BRKINT = 0x00000002;
-        IGNPAR = 0x00000004;
         PARMRK = 0x00000008;
         INLCR = 0x00000040;
         IGNCR = 0x00000080;
         ICRNL = 0x00000100;
-        ECHONL = 0x00000040;
-        IEXTEN = 0x00008000;
-        CLOCAL = 0x00000800;
+        ECHONL = 0x00000010;
+        IEXTEN = 0x00000400;
+        CLOCAL = 0x00008000;
         OPOST = 0x00000001;
-        VSTART = 0x00000008;
+        VSTART = 0x0000000C;
         TCSANOW = 0x00000000;
-        VSTOP = 0x00000009;
-        VMIN = 0x00000006;
-        VTIME = 0x00000005;
-        VEOF = 0x00000004;
-        TIOCMGET = 0x00005415;
+        VSTOP = 0x0000000D;
+        VMIN = 0x00000010;
+        VTIME = 0x00000011;
+        VEOF = 0x00000000;
+        TIOCMGET = 0x4004746A;
         TIOCM_CTS = 0x00000020;
         TIOCM_DSR = 0x00000100;
         TIOCM_RI = 0x00000080;
         TIOCM_CD = 0x00000040;
         TIOCM_DTR = 0x00000002;
         TIOCM_RTS = 0x00000004;
-        ICANON = 0x00000002;
+        ICANON = 0x00000100;
         ECHO = 0x00000008;
-        ECHOE = 0x00000010;
-        ISIG = 0x00000001;
-        TIOCMSET = 0x00005418;
-        IXON = 0x00000400;
-        IXOFF = 0x00001000;
+        ECHOE = 0x00000002;
+        ISIG = 0x00000080;
+        TIOCMSET = 0x8004746D;
+        IXON = 0x00000200;
+        IXOFF = 0x00000400;
         IXANY = 0x00000800;
-        CRTSCTS = 0x80000000;
+        CRTSCTS = 0x00030000;
         TCSADRAIN = 0x00000001;
         INPCK = 0x00000010;
         ISTRIP = 0x00000020;
-        CSIZE = 0x00000030;
-        TCIFLUSH = 0x00000000;
-        TCOFLUSH = 0x00000001;
-        TCIOFLUSH = 0x00000002;
+        CSIZE = 0x00000300;
+        TCIFLUSH = 0x00000001;
+        TCOFLUSH = 0x00000002;
+        TCIOFLUSH = 0x00000003;
         CS5 = 0x00000000;
-        CS6 = 0x00000010;
-        CS7 = 0x00000020;
-        CS8 = 0x00000030;
-        CSTOPB = 0x00000040;
-        CREAD = 0x00000080;
-        PARENB = 0x00000100;
-        PARODD = 0x00000200;
+        CS6 = 0x00000100;
+        CS7 = 0x00000200;
+        CS8 = 0x00000300;
+        CSTOPB = 0x00000400;
+        CREAD = 0x00000800;
+        PARENB = 0x00001000;
+        PARODD = 0x00002000;
         B0 = 0;
-        B50 = 1;
-        B75 = 2;
-        B110 = 3;
-        B134 = 4;
-        B150 = 5;
-        B200 = 6;
-        B300 = 7;
-        B600 = 8;
-        B1200 = 9;
-        B1800 = 10;
-        B2400 = 11;
-        B4800 = 12;
-        B9600 = 13;
-        B19200 = 14;
-        B38400 = 15;
-        B57600 = 4097;
-        B115200 = 4098;
-        B230400 = 4099;
-        //poll.h stuff
+        B50 = 50;
+        B75 = 75;
+        B110 = 110;
+        B134 = 134;
+        B150 = 150;
+        B200 = 200;
+        B300 = 300;
+        B600 = 600;
+        B1200 = 600;
+        B1800 = 1800;
+        B2400 = 2400;
+        B4800 = 4800;
+        B9600 = 9600;
+        B19200 = 19200;
+        B38400 = 38400;
+        B7200 = 7200;
+        B14400 = 14400;
+        B28800 = 28800;
+        B57600 = 57600;
+        B76800 = 76800;
+        B115200 = 115200;
+        B230400 = 230400;
+        // poll.h stuff
         POLLIN = 0x0001;
         POLLPRI = 0x0002;
         POLLOUT = 0x0004;
         POLLERR = 0x0008;
         POLLNVAL = 0x0020;
+        // select.h stuff
     }
 
     public int errno() {
@@ -502,23 +404,23 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
     }
 
     public int cfgetispeed(Termios termios) {
-        return m_Clib.cfgetispeed(new termios(termios));
+        return m_Clib.cfgetispeed(new termios(termios)).intValue();
     }
 
     public int cfgetospeed(Termios termios) {
-        return m_Clib.cfgetospeed(new termios(termios));
+        return m_Clib.cfgetospeed(new termios(termios)).intValue();
     }
 
     public int cfsetispeed(Termios termios, int speed) {
         termios t = new termios(termios);
-        int ret = m_Clib.cfsetispeed(t, speed);
+        int ret = m_Clib.cfsetispeed(t, new NativeLong(speed));
         t.update(termios);
         return ret;
     }
 
     public int cfsetospeed(Termios termios, int speed) {
         termios t = new termios(termios);
-        int ret = m_Clib.cfsetospeed(t, speed);
+        int ret = m_Clib.cfsetospeed(t, new NativeLong(speed));
         t.update(termios);
         return ret;
     }
@@ -573,7 +475,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
             tout = new timeval(timeout);
         }
 
-        return m_ClibND.select(nfds, (fd_set) rfds, (fd_set) wfds, (fd_set) efds, tout);
+        return m_Clib.select(nfds, (fd_set) rfds, (fd_set) wfds, (fd_set) efds, tout);
     }
 
     public int poll(Pollfd fds[], int nfds, int timeout) {
@@ -602,74 +504,8 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
     }
 
     public int ioctl(int fd, int cmd, int... data) {
+        // At this time, all ioctl commands we have defined are either no parameter or 4 byte parameter.
         return m_Clib.ioctl(fd, cmd, data);
-    }
-
-    // This ioctl is Linux specific, so keep it private for now
-    private int ioctl(int fd, int cmd, serial_struct data) {
-        // Do the logging here as this does not go through the JTermios which normally does the logging
-        log = log && log(5, "> ioctl(%d,%d,%s)\n", fd, cmd, data);
-        int ret = m_Clib.ioctl(fd, cmd, data);
-        log = log && log(3, "< tcsetattr(%d,%d,%s) => %d\n", fd, cmd, data, ret);
-        return ret;
-    }
-
-    public String getPortNamePattern() {
-        // First we have to determine which serial drivers exist and which
-        // prefixes they use
-        final List<String> prefixes = new ArrayList<String>();
-
-        try {
-            BufferedReader drivers = new BufferedReader(new InputStreamReader(new FileInputStream("/proc/tty/drivers"), "US-ASCII"));
-            String line;
-            while ((line = drivers.readLine()) != null) {
-                // /proc/tty/drivers contains the prefix in the second column
-                // and "serial" in the fifth
-
-                String[] parts = line.split(" +");
-                if (parts.length != 5) {
-                    continue;
-                }
-
-                if (!"serial".equals(parts[4])) {
-                    continue;
-                }
-
-                // Sanity check the prefix
-                if (!parts[1].startsWith("/dev/")) {
-                    continue;
-                }
-
-                prefixes.add(parts[1].substring(5));
-            }
-            drivers.close();
-        } catch (IOException e) {
-            log = log && log(1, "failed to read /proc/tty/drivers\n");
-
-            prefixes.add("ttyS");
-            prefixes.add("ttyUSB");
-            prefixes.add("ttyACM");
-        }
-
-        // Now build the pattern from the known prefixes
-        StringBuilder pattern = new StringBuilder();
-
-        pattern.append('^');
-
-        boolean first = true;
-        for (String prefix : prefixes) {
-            if (first) {
-                first = false;
-            } else {
-                pattern.append('|');
-            }
-
-            pattern.append("(");
-            pattern.append(prefix);
-            pattern.append(".+)");
-        }
-
-        return pattern.toString();
     }
 
     public List<String> getPortList() {
@@ -680,16 +516,16 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         }
         String[] devs = dir.list();
         LinkedList<String> list = new LinkedList<String>();
-
-        Pattern p = JTermios.getPortNamePattern(this);
         if (devs != null) {
             for (int i = 0; i < devs.length; i++) {
                 String s = devs[i];
-                if (p.matcher(s).matches()) {
+                if (s.startsWith("cua") || s.startsWith("tty")) {
                     list.add(s);
                 }
             }
+
         }
+
         return list;
     }
 
@@ -697,82 +533,24 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
     }
 
+    public String getPortNamePattern() {
+        return "^(tty\\.|cu\\.).*";
+    }
+
     public int setspeed(int fd, Termios termios, int speed) {
-        int c = speed;
         int r;
-        for (int i = 0; i < m_BaudRates.length; i += 2) {
-            if (m_BaudRates[i] == speed) {
-
-                // found the baudrate from the table
-                // just in case custom divisor was in use, try to turn it off first
-                serial_struct ss = new serial_struct();
-
-                r = ioctl(fd, TIOCGSERIAL, ss);
-                if (r == 0) {
-                    ss.flags &= ~ASYNC_SPD_MASK;
-                    r = ioctl(fd, TIOCSSERIAL, ss);
-                }
-
-                // now set the speed with the constant from the table
-                c = m_BaudRates[i + 1];
-                if ((r = JTermios.cfsetispeed(termios, c)) != 0) {
-                    return r;
-                }
-                if ((r = JTermios.cfsetospeed(termios, c)) != 0) {
-                    return r;
-                }
-                if ((r = JTermios.tcsetattr(fd, TCSANOW, termios)) != 0) {
-                    return r;
-                }
-
-                return 0;
-            }
+        r = cfsetispeed(termios, speed);
+        if (r == 0) {
+            r = cfsetospeed(termios, speed);
         }
-
-        // baudrate not defined in the table, try custom divisor approach
-        // configure port to use custom speed instead of 38400
-        serial_struct ss = new serial_struct();
-        if ((r = ioctl(fd, TIOCGSERIAL, ss)) != 0) {
-            return r;
+        if (r == 0) {
+            r = tcsetattr(fd, TCSANOW, termios);
         }
-        ss.flags = (ss.flags & ~ASYNC_SPD_MASK) | ASYNC_SPD_CUST;
-
-        if (speed == 0) {
-            log = log && log(1, "unable to set custom baudrate %d \n", speed);
-            return -1;
-        }
-
-        ss.custom_divisor = (ss.baud_base + (speed / 2)) / speed;
-
-        if (ss.custom_divisor == 0) {
-            log = log && log(1, "unable to set custom baudrate %d (possible division by zero)\n", speed);
-            return -1;
-        }
-
-        int closestSpeed = ss.baud_base / ss.custom_divisor;
-
-        if (closestSpeed < speed * 98 / 100 || closestSpeed > speed * 102 / 100) {
-            log = log && log(1, "best available baudrate %d not close enough to requested %d \n", closestSpeed, speed);
-            return -1;
-        }
-
-        if ((r = ioctl(fd, TIOCSSERIAL, ss)) != 0) {
-            return r;
-        }
-
-        if ((r = JTermios.cfsetispeed(termios, B38400)) != 0) {
-            return r;
-        }
-        if ((r = JTermios.cfsetospeed(termios, B38400)) != 0) {
-            return r;
-        }
-        if ((r = JTermios.tcsetattr(fd, TCSANOW, termios)) != 0) {
-            return r;
-        }
-        return 0;
+        return r;
     }
 
     public int pipe(int[] fds) {
         return m_Clib.pipe(fds);
     }
+
 }
