+++ src/main/java/com/sparetimelabs/serial/termios/impl/MacOSXTermios.java
@@ -27,7 +27,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
  */
-package com.sparetimelabs.serial.termios.solaris;
+package com.sparetimelabs.serial.termios.impl;
 
 import com.sun.jna.*;
 import java.io.File;
@@ -42,16 +42,17 @@ import com.sparetimelabs.serial.termios.TimeVal;
 import static com.sparetimelabs.serial.termios.JTermios.*;
 import static com.sparetimelabs.serial.termios.JTermios.JTermiosLogging.log;
 
-public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.JTermiosInterface {
+public class MacOSXTermios implements com.sparetimelabs.serial.termios.JTermios.JTermiosInterface {
 
+    private static int IOSSIOSPEED = 0x80045402;
     private static String DEVICE_DIR_PATH = "/dev/";
     static C_lib_DirectMapping m_ClibDM;
     static C_lib m_Clib;
-    static NonDirectCLib m_ClibND;
+//	static NonDirectCLib m_ClibND;
 
     static {
         Native.setPreserveLastError(true);
-        m_ClibND = (NonDirectCLib) Native.loadLibrary(Platform.C_LIBRARY_NAME, NonDirectCLib.class);
+//		m_ClibND = (NonDirectCLib) Native.loadLibrary(Platform.C_LIBRARY_NAME, NonDirectCLib.class);
         Native.register(C_lib_DirectMapping.class, NativeLibrary.getInstance(Platform.C_LIBRARY_NAME));
         m_ClibDM = new C_lib_DirectMapping();
         m_Clib = m_ClibDM;
@@ -67,7 +68,11 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         native public int fcntl(int fd, int cmd, int arg);
 
-        native public int ioctl(int fd, int cmd, int[] arg);
+        native public int ioctl(int fd, NativeLong cmd);
+
+        native public int ioctl(int fd, NativeLong cmd, int arg);
+
+        native public int ioctl(int fd, NativeLong cmd, int[] arg);
 
         native public int open(String path, int flags);
 
@@ -94,6 +99,9 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         native public void perror(String msg);
 
         native public int tcsendbreak(int fd, int duration);
+
+        native public int select(int n, fd_set read, fd_set write, fd_set error, timeval timeout);
+
     }
 
     public interface C_lib extends com.sun.jna.Library {
@@ -106,7 +114,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public int fcntl(int fd, int cmd, int arg);
 
-        public int ioctl(int fd, int cmd, int[] arg);
+        public int ioctl(int fd, NativeLong cmd, int[] arg);
 
         public int open(String path, int flags);
 
@@ -134,15 +142,16 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public int tcsendbreak(int fd, int duration);
 
-    }
-
-    public interface NonDirectCLib extends com.sun.jna.Library {
-
         public int select(int n, fd_set read, fd_set write, fd_set error, timeval timeout);
 
-        public int poll(pollfd.ByReference pfds, int nfds, int timeout);
     }
 
+//	public interface NonDirectCLib extends com.sun.jna.Library {
+//
+//		public int ioctl(int fd, NativeLong cmd, NativeLong[] arg);
+//
+//		//       public int poll(pollfd.ByReference pfds, int nfds, int timeout);
+//	}
     static public class timeval extends Structure {
 
         public NativeLong tv_sec;
@@ -166,6 +175,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
         public static class ByReference extends pollfd implements Structure.ByReference {
         }
+
         public int fd;
         public short events;
         public short revents;
@@ -191,15 +201,9 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
     static public class fd_set extends Structure implements FDSet {
 
-        private final static int NFBBITS = NativeLong.SIZE * 8;
+        private final static int NFBBITS = 32;
         private final static int fd_count = 1024;
-        public NativeLong[] fd_array = new NativeLong[(fd_count + NFBBITS - 1) / NFBBITS];
-
-        public fd_set() {
-            for (int i = 0; i < fd_array.length; ++i) {
-                fd_array[i] = new NativeLong();
-            }
-        }
+        public int[] fd_array = new int[(fd_count + NFBBITS - 1) / NFBBITS];
 
         @Override
         protected List getFieldOrder() {
@@ -209,32 +213,32 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         }
 
         public void FD_SET(int fd) {
-            fd_array[fd / NFBBITS].setValue(fd_array[fd / NFBBITS].longValue() | (1L << (fd % NFBBITS)));
+            fd_array[fd / NFBBITS] |= (1 << (fd % NFBBITS));
         }
 
         public boolean FD_ISSET(int fd) {
-            return (fd_array[fd / NFBBITS].longValue() & (1L << (fd % NFBBITS))) != 0;
+            return (fd_array[fd / NFBBITS] & (1 << (fd % NFBBITS))) != 0;
         }
 
         public void FD_ZERO() {
-            for (NativeLong fd : fd_array) {
-                fd.setValue(0L);
-            }
+            Arrays.fill(fd_array, 0);
         }
 
         public void FD_CLR(int fd) {
-            fd_array[fd / NFBBITS].setValue(fd_array[fd / NFBBITS].longValue() & ~(1L << (fd % NFBBITS)));
+            fd_array[fd / NFBBITS] &= ~(1 << (fd % NFBBITS));
         }
 
     }
 
     static public class termios extends Structure {
 
-        public int c_iflag;
-        public int c_oflag;
-        public int c_cflag;
-        public int c_lflag;
-        public byte[] c_cc = new byte[32];
+        public NativeLong c_iflag;
+        public NativeLong c_oflag;
+        public NativeLong c_cflag;
+        public NativeLong c_lflag;
+        public byte[] c_cc = new byte[20];
+        public NativeLong c_ispeed;
+        public NativeLong c_ospeed;
 
         @Override
         protected List getFieldOrder() {
@@ -243,7 +247,9 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
                     "c_oflag",//
                     "c_cflag",//
                     "c_lflag",//
-                    "c_cc"//
+                    "c_cc",//
+                    "c_ispeed",//
+                    "c_ospeed"//
             );
         }
 
@@ -251,138 +257,28 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         }
 
         public termios(com.sparetimelabs.serial.termios.Termios t) {
-            c_iflag = t.c_iflag;
-            c_oflag = t.c_oflag;
-            c_cflag = t.c_cflag;
-            c_lflag = t.c_lflag;
+            c_iflag.setValue(t.c_iflag);
+            c_oflag.setValue(t.c_oflag);
+            c_cflag.setValue(t.c_cflag);
+            c_lflag.setValue(t.c_lflag);
             System.arraycopy(t.c_cc, 0, c_cc, 0, Math.min(t.c_cc.length, c_cc.length));
+            c_ispeed.setValue(t.c_ispeed);
+            c_ospeed.setValue(t.c_ospeed);
         }
 
         public void update(com.sparetimelabs.serial.termios.Termios t) {
-            t.c_iflag = c_iflag;
-            t.c_oflag = c_oflag;
-            t.c_cflag = c_cflag;
-            t.c_lflag = c_lflag;
+            t.c_iflag = c_iflag.intValue();
+            t.c_oflag = c_oflag.intValue();
+            t.c_cflag = c_cflag.intValue();
+            t.c_lflag = c_lflag.intValue();
             System.arraycopy(c_cc, 0, t.c_cc, 0, Math.min(t.c_cc.length, c_cc.length));
+            t.c_ispeed = c_ispeed.intValue();
+            t.c_ospeed = c_ospeed.intValue();
         }
     }
 
-    public JTermiosImpl() {
+    public MacOSXTermios() {
         log = log && log(1, "instantiating %s\n", getClass().getCanonicalName());
-
-        // sys/filio.h stuff
-        FIONREAD = 0x4004667F;
-
-        //fcntl.h stuff
-        O_RDWR = 0x00000002;
-        O_NONBLOCK = 0x00000080;
-        O_NOCTTY = 0x00000800;
-        O_NDELAY = 0x00000004;
-        F_GETFL = 0x00000003;
-        F_SETFL = 0x00000004;
-
-        //errno.h stuff
-        EAGAIN = 11;
-        EBADF = 9;
-        EACCES = 22;
-        EEXIST = 17;
-        EINTR = 4;
-        EINVAL = 22;
-        EIO = 5;
-        EISDIR = 21;
-        ELOOP = 90;
-        EMFILE = 24;
-        ENAMETOOLONG = 78;
-        ENFILE = 23;
-        ENOENT = 2;
-        ENOSR = 63;
-        ENOSPC = 28;
-        ENOTDIR = 20;
-        ENXIO = 6;
-        EOVERFLOW = 79;
-        EROFS = 30;
-        ENOTSUP = 48;
-
-        //termios.h stuff
-        TIOCM_RNG = 0x00000080;
-        TIOCM_CAR = 0x00000040;
-        IGNBRK = 0x00000001;
-        BRKINT = 0x00000002;
-        IGNPAR = 0x00000004;
-        PARMRK = 0x00000008;
-        INLCR = 0x00000040;
-        IGNCR = 0x00000080;
-        ICRNL = 0x00000100;
-        ECHONL = 0x00000040;
-        IEXTEN = 0x00008000;
-        CLOCAL = 0x00000800;
-        OPOST = 0x00000001;
-        VSTART = 0x00000008;
-        TCSANOW = 0x0000540E;
-        VSTOP = 0x00000009;
-        VMIN = 0x00000004;
-        VTIME = 0x00000005;
-        VEOF = 0x00000004;
-        TIOCMGET = 0x0000741D;
-        TIOCM_CTS = 0x00000020;
-        TIOCM_DSR = 0x00000100;
-        TIOCM_RI = 0x00000080;
-        TIOCM_CD = 0x00000040;
-        TIOCM_DTR = 0x00000002;
-        TIOCM_RTS = 0x00000004;
-        ICANON = 0x00000002;
-        ECHO = 0x00000008;
-        ECHOE = 0x00000010;
-        ISIG = 0x00000001;
-        TIOCMSET = 0x0000741A;
-        IXON = 0x00000400;
-        IXOFF = 0x00001000;
-        IXANY = 0x00000800;
-        CRTSCTS = 0x80000000;
-        TCSADRAIN = 0x0000540F;
-        INPCK = 0x00000010;
-        ISTRIP = 0x00000020;
-        CSIZE = 0x00000030;
-        TCIFLUSH = 0x00000000;
-        TCOFLUSH = 0x00000001;
-        TCIOFLUSH = 0x00000002;
-        CS5 = 0x00000000;
-        CS6 = 0x00000010;
-        CS7 = 0x00000020;
-        CS8 = 0x00000030;
-        CSTOPB = 0x00000040;
-        CREAD = 0x00000080;
-        PARENB = 0x00000100;
-        PARODD = 0x00000200;
-        B0 = 0;
-        B50 = 1;
-        B75 = 2;
-        B110 = 3;
-        B134 = 4;
-        B150 = 5;
-        B200 = 6;
-        B300 = 7;
-        B600 = 8;
-        B1200 = 8;
-        B1800 = 10;
-        B2400 = 11;
-        B4800 = 12;
-        B9600 = 13;
-        B19200 = 14;
-        B38400 = 15;
-        B57600 = 16;
-        B76800 = 17;
-        B115200 = 18;
-        B230400 = 20;
-
-        //poll.h stuff
-        POLLIN = 0x0001;
-        POLLPRI = 0x0002;
-        POLLOUT = 0x0004;
-        POLLERR = 0x0008;
-        POLLNVAL = 0x0020;
-
-        //select.h stuff
     }
 
     public int errno() {
@@ -475,28 +371,15 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
             tout = new timeval(timeout);
         }
 
-        return m_ClibND.select(nfds, (fd_set) rfds, (fd_set) wfds, (fd_set) efds, tout);
+        return m_Clib.select(nfds, (fd_set) rfds, (fd_set) wfds, (fd_set) efds, tout);
     }
 
     public int poll(Pollfd fds[], int nfds, int timeout) {
-        if (nfds <= 0 || nfds > fds.length) {
-            throw new java.lang.IllegalArgumentException("nfds " + nfds + " must be <= fds.length " + fds.length);
-        }
-        pollfd.ByReference parampfds = new pollfd.ByReference();
-        pollfd[] pfds = (pollfd[]) parampfds.toArray(nfds);
-        for (int i = 0; i < nfds; i++) {
-            pfds[i].fd = fds[i].fd;
-            pfds[i].events = fds[i].events;
-        }
-        int ret = m_ClibND.poll(parampfds, nfds, timeout);
-        for (int i = 0; i < nfds; i++) {
-            fds[i].revents = pfds[i].revents;
-        }
-        return ret;
+        throw new UnsupportedOperationException("Poll not supported");
     }
 
     public boolean canPoll() {
-        return true;
+        return false;
     }
 
     public FDSet newFDSet() {
@@ -505,11 +388,7 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
 
     public int ioctl(int fd, int cmd, int... data) {
         // At this time, all ioctl commands we have defined are either no parameter or 4 byte parameter.
-        return m_Clib.ioctl(fd, cmd, data);
-    }
-
-    public String getPortNamePattern() {
-        return ".*";
+        return m_Clib.ioctl(fd, new NativeLong(0xFFFFFFFFL & cmd), data);
     }
 
     public List<String> getPortList() {
@@ -533,94 +412,33 @@ public class JTermiosImpl implements com.sparetimelabs.serial.termios.JTermios.J
         return list;
     }
 
+    public String getPortNamePattern() {
+        return "^(tty\\.|cu\\.).*";
+    }
+
     public void shutDown() {
+
     }
 
     public int setspeed(int fd, Termios termios, int speed) {
-        int br = speed;
-        switch (speed) {
-            case 50:
-                br = B50;
-                break;
-            case 75:
-                br = B75;
-                break;
-            case 110:
-                br = B110;
-                break;
-            case 134:
-                br = B134;
-                break;
-            case 150:
-                br = B150;
-                break;
-            case 200:
-                br = B200;
-                break;
-            case 300:
-                br = B300;
-                break;
-            case 600:
-                br = B600;
-                break;
-            case 1200:
-                br = B1200;
-                break;
-            case 1800:
-                br = B1800;
-                break;
-            case 2400:
-                br = B2400;
-                break;
-            case 4800:
-                br = B4800;
-                break;
-            case 9600:
-                br = B9600;
-                break;
-            case 19200:
-                br = B19200;
-                break;
-            case 38400:
-                br = B38400;
-                break;
-            case 7200:
-                br = B7200;
-                break;
-            case 14400:
-                br = B14400;
-                break;
-            case 28800:
-                br = B28800;
-                break;
-            case 57600:
-                br = B57600;
-                break;
-            case 76800:
-                br = B76800;
-                break;
-            case 115200:
-                br = B115200;
-                break;
-            case 230400:
-                br = B230400;
-                break;
-        }
         int r;
-        if ((r = cfsetispeed(termios, br)) != 0) {
-            return r;
+        r = cfsetispeed(termios, speed);
+        if (r == 0) {
+            r = cfsetospeed(termios, speed);
         }
-        if ((r = cfsetospeed(termios, br)) != 0) {
-            return r;
+        if (r == 0) {
+            r = tcsetattr(fd, TCSANOW, termios);
         }
-        if ((r = tcsetattr(fd, TCSANOW, termios)) != 0) {
-            return r;
+        if (r != 0) {
+            // Darell Tan had patched RXTX with this sequence, so lets try this
+            if (cfsetispeed(termios, B9600) == 0 && cfsetospeed(termios, B9600) == 0 && tcsetattr(fd, TCSANOW, termios) == 0) {
+                r = ioctl(fd, IOSSIOSPEED, speed);
+            }
         }
-        return 0;
+        return r;
     }
 
     public int pipe(int[] fds) {
         return m_Clib.pipe(fds);
     }
-
 }
